To avoid package naming conflicts, rename inquirery runner module and package from just "inquirer" to "inquirer".
This way it's also consistent with other runner names.
