--- contrib/runners/inquirer/runner.yaml
@@ -1,1 +1,0 @@
-./inquirer/runner.yaml
