+++ contrib/runners/inquirer_runner/MANIFEST.in
