+++ contrib/runners/inquirer_runner/runner.yaml
@@ -1,0 +1,1 @@
+inquirer_runner/runner.yaml
