+++ st2api/st2api/controllers/exp/inquiries.py
@@ -191,7 +191,7 @@ class InquiriesController(ResourceController):
         :rtype: str - Error message, if any
         """
 
-        if inquiry_candidate.runner.get('runner_module') != "inquirer":
+        if inquiry_candidate.runner.get('name') != 'inquirer':
             return (False, '%s is not an Inquiry.' % inquiry_candidate.id)
 
         if inquiry_candidate.status == action_constants.LIVEACTION_STATUS_TIMED_OUT:
