--- st2tests/st2tests/fixtures/packs/runners/inquirer
@@ -1,1 +1,0 @@
-../../../../../contrib/runners/inquirer
