+++ st2tests/st2tests/fixtures/packs/runners/inquirer_runner
@@ -1,0 +1,1 @@
+../../../../../contrib/runners/inquirer_runner
